made seperate file for each ingredients. Added function for each ingredients.
